requirements.txt
django
djangorestframework
pydantic
requests